Refactor to <Lyrics>
